Blog section update

+++ src/pages/Page.tsx
@@ -1,40 +1,59 @@
 import { Fragment, useEffect, useState } from 'react';
 import DynamicComponent from '../components/Component';
 import { useParams } from 'react-router-dom';
-import { AiFillEdit, AiFillDelete, AiOutlineAppstoreAdd } from 'react-icons/ai';
+import { AiFillEdit, AiFillDelete } from 'react-icons/ai';
+import { MdAddCard } from 'react-icons/md';
 import { deleteComponentOfPage } from '../firebase/util';
 import { CModal } from '../components/modal/Modal';
 import { ListGroup } from 'react-bootstrap';
 import NotFound from './NotFound';
 import { getMockdata } from '../util/mockData.util';
 import { subscribePageComponents } from '../firebase/firebase.util';
+import { MdOutlineRepeatOne } from 'react-icons/md';
 // import { scrollTo } from 'scroll-js';
 
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
 export default function Page(props: any) {
   const { isAuth, editable } = props;
   const [data, setData] = useState([]);
   const [modalData, setModalData] = useState(null);
   const [modalAction, setModalAction] = useState(null);
   const [modalState, setModalState] = useState(false);
-  const { page } = useParams();
+  const [docId, setDocId] = useState(null);
+  const { page: collection } = useParams();
   const [loading, setLoading] = useState(true);
 
   const handleClose = () => setModalState(false);
 
-  const onDelete = (docId, collection) => {
-    if (window.confirm('Are you sure you want to Delete?'))
+  const onDelete = () => {
+    if (!docId) return;
+    if (window.confirm('Are you sure you want to Delete?')) {
       deleteComponentOfPage(collection, docId, console.log);
+    }
   };
 
-  const onEdit = (docId, collection, data) => {
+  const onEdit = () => {
+    if (!docId) return;
     setModalState(true);
     setModalAction('EDIT');
-    const patched = getMockdata(data.component, data);
-    setModalData({ docId, collection, data: patched });
+    const { data: modalData } = data.find((f) => f.id === docId);
+    setModalData({ docId, collection, data: getMockdata(modalData.component, modalData) });
+  };
+
+  const componentClick = (id: string) => {
+    if (isAuth && editable === 'true') {
+      setDocId(id);
+    }
   };
 
   const onAddComponent = () => {
     setModalAction('ADD');
+    setModalState(true);
+    setModalData(null);
+  };
+
+  const onOrder = (): void => {
+    setModalAction('ORDER');
     setModalState(true);
     setModalData(null);
   };
@@ -50,7 +69,7 @@ export default function Page(props: any) {
 
     let unsubscribe;
     const init = async () => {
-      unsubscribe = await subscribePageComponents(page, (res) => {
+      unsubscribe = await subscribePageComponents(collection, (res) => {
         setData(res);
       });
     };
@@ -60,10 +79,9 @@ export default function Page(props: any) {
       if (unsubscribe) unsubscribe();
       window.removeEventListener('load', () => setLoading(false));
     };
-  }, [page]);
+  }, [collection]);
 
   if (loading) {
-    // console.log('Loading')
     return (
       <div style={{ position: 'absolute', inset: 0, background: 'black' }}>
         <div className="animate-spin rounded-full h-12 w-12 border-t-4 border-b-4 border-blue-500"></div>
@@ -77,22 +95,16 @@ export default function Page(props: any) {
       {data
         ?.sort((a, b) => a.data.order - b.data.order)
         .map((m) => (
-          <div key={m.id} style={{ position: 'relative' }}>
+          <div
+            className={
+              `${isAuth && editable === 'true' ? 'pointer' : ''} ` +
+              `${docId === m.id ? 'highlight' : ''}`
+            }
+            key={m.id}
+            style={{ position: 'relative' }}
+            onClick={() => componentClick(m.id)}
+          >
             <DynamicComponent data={m.data} id={m.id}></DynamicComponent>
-            {isAuth && editable === 'true' ? (
-              <div className="h5 editable">
-                <ListGroup horizontal className="pointer">
-                  <ListGroup.Item onClick={() => onEdit(m.id, page, m.data)}>
-                    <AiFillEdit />
-                  </ListGroup.Item>
-                  <ListGroup.Item onClick={() => onDelete(m.id, page)}>
-                    <AiFillDelete />
-                  </ListGroup.Item>
-                </ListGroup>
-              </div>
-            ) : (
-              ''
-            )}
           </div>
         ))}
       {modalState ? (
@@ -101,7 +113,7 @@ export default function Page(props: any) {
           close={handleClose}
           data={modalData}
           action={modalAction}
-          pageID={page}
+          pageID={collection}
         ></CModal>
       ) : (
         ''
@@ -109,12 +121,21 @@ export default function Page(props: any) {
       {isAuth && editable === 'true' && (
         <div className="floating-icon">
           <ListGroup
-            horizontal
+            horizontal="false"
             className="pointer"
             style={{ borderTopRightRadius: 0, borderBottomRightRadius: 0 }}
           >
             <ListGroup.Item onClick={onAddComponent}>
-              <AiOutlineAppstoreAdd />
+              <MdAddCard />
+            </ListGroup.Item>
+            <ListGroup.Item className={!docId ? 'disabled' : ''} onClick={onEdit}>
+              <AiFillEdit />
+            </ListGroup.Item>
+            <ListGroup.Item className={!docId ? 'disabled' : ''} onClick={onDelete}>
+              <AiFillDelete />
+            </ListGroup.Item>
+            <ListGroup.Item onClick={onOrder}>
+              <MdOutlineRepeatOne />
             </ListGroup.Item>
           </ListGroup>
         </div>
@@ -122,7 +143,7 @@ export default function Page(props: any) {
     </Fragment>
   ) : (
     <Fragment>
-      <NotFound pageID={page}></NotFound>
+      <NotFound pageID={collection}></NotFound>
     </Fragment>
   );
 }
